modify the pages structure
