+++ frontend/src/auth/context/utils.js
@@ -4,7 +4,7 @@ import axios from 'src/utils/axios';
 
 // ----------------------------------------------------------------------
 
-function jwtDecode(token) {
+export function jwtDecode(token) {
   const base64Url = token.split('.')[1];
   const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
   const jsonPayload = decodeURIComponent(
