+++ frontend/src/pages/instructor/course-create/sections/FirstStep.jsx
