+++ frontend/src/pages/instructor/course-create/sections/SecondStep.jsx
