+++ frontend/src/pages/instructor/course-create/sections/ThirdStep.jsx
