+++ frontend/src/pages/instructor/course-info/sections/Enrollments.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
 import { Helmet } from "react-helmet-async";
-import { paths } from "../../../routes/paths";
+import { paths } from "src/routes/paths";
 const Enrollments = ({ data }) => {
   const navigate = useNavigate();
